backend/requirements.txt
Django==5.2.2
PyMySQL==1.1.1
djangorestframework==3.16.0
django-cors-headers==4.7.0
Pillow==11.2.1
